% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_start_points.R
\name{add_start_points}
\alias{add_start_points}
\title{Add Start Points}
\usage{
add_start_points(final_routes)
}
\arguments{
\item{final_routes}{the sf dataframe with hiking routes. Created by function "individual_routes".}
}
\value{
Returns a the sf dataframe "final_routes" with the added starting points and their lat and long in 3 new columns.
}
\description{
\code{add_start_points} Adds starting points and their lat and long to sf df final_routes.
}
\examples{
\dontrun{
library(hike4u)

final_routes <- add_start_points(final_routes)
}

}
