% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_closeness.R
\name{calculate_closeness}
\alias{calculate_closeness}
\title{Calculate Closeness}
\usage{
calculate_closeness(final_routes, poi)
}
\arguments{
\item{final_routes}{the sf dataframe with hiking routes. Created by function "add_start_points".}

\item{poi}{the POI as an sf dataframe with your location as a point. Created by function "create_poi_df".}
}
\value{
Returns a the sf dataframe "final_routes" with 3 new columns: the closeness value to your location, the distance between the route start point and your location, the length of the route.
}
\description{
\code{calculate_closeness} Calculates distances between each starting point in final_routes and the POI. Ranks these distances to determine the closeness of each. Calculates length of route.
}
\examples{
\dontrun{
library(hike4u)

final_routes <- calculate_closeness(final_routes, poi)
}

}
