% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ext.R
\name{calculate_ext}
\alias{calculate_ext}
\title{Calculate Extent}
\usage{
calculate_ext(final_routes, poi, closeness_value)
}
\arguments{
\item{final_routes}{the sf dataframe with hiking routes. Created by function "calculate_closeness".}

\item{poi}{the POI as an sf dataframe with your location as a point. Created by function "create_poi_df".}

\item{closeness_value}{the defined closeness value as a Value. Created by function "def_closeness".}
}
\value{
Returns an sf dataframe "ext" with a polygon of the extent. Or "No rows found with the specified closeness value."
}
\description{
\code{calculate_ext} Calculates the extent the map should have depending on which route is chosen by closeness value.
}
\examples{
\dontrun{
library(hike4u)

ext <- calculate_ext(final_routes, poi, closeness_value)
}

}
