% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_buffer.R
\name{create_buffer}
\alias{create_buffer}
\title{Create Buffer Value}
\usage{
create_buffer(buffer)
}
\arguments{
\item{buffer}{the defined buffer around the POI in meters.}
}
\value{
Returns a number value "buffer" with the size in meters.
}
\description{
\code{create_buffer} Creates a number value with your defined buffer size.
}
\examples{
\dontrun{
library(hike4u)

# Assign 10000 meters (10 km) to buffer
buffer <- create_buffer(10000)
}

}
