% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_poi_df.R
\name{create_poi_df}
\alias{create_poi_df}
\title{Create POI Dataframe}
\usage{
create_poi_df(longitude, latitude)
}
\arguments{
\item{longitude}{the longitude of your location.}

\item{latitude}{the latitude of your location.}
}
\value{
Returns an sf dataframe "poi" with your location as a point in "WGS 84" "EPSG:4326" projection.
}
\description{
\code{create_poi_df} Creates an sf dataframe with your location as a point.
}
\examples{
\dontrun{
library(hike4u)

poi <- create_poi_df(longitude = 9.93389691622025, latitude = 49.79895823510417)
}

}
