% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_closeness.R
\name{def_closeness}
\alias{def_closeness}
\title{Define Closeness Value}
\usage{
def_closeness(closeness_value)
}
\arguments{
\item{closeness_value}{the defined closeness value. Choose between 1 and the max number of routes in your area. 1 = closest route to your location.}
}
\value{
Returns a number value "buffer" with the size in meters.
}
\description{
\code{def_closeness} Creates a number value with your defined closeness value. Choose 1 for the first run of the package to then get returned which values you could choose as max.
}
\examples{
\dontrun{
library(hike4u)

# Assign a closeness value of 1 to choose the closest route to your location (POI)
closeness_value <- def_closeness(1)
}

}
