% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hiking_routes.R
\name{get_hiking_routes}
\alias{get_hiking_routes}
\title{Get Hiking Routes}
\usage{
get_hiking_routes(poi, buffer)
}
\arguments{
\item{poi}{the POI as an sf dataframe with your location as a point. Created by function "create_poi_df".}

\item{buffer}{the defined buffer around the POI in meters as a Value. Created by function "create_buffer".}
}
\value{
Returns an sf dataframe "hiking_routes" with the relevant hiking routes. If no hiking routes are found "No hiking routes found within your buffer area." is printed.
}
\description{
\code{get_hiking_routes} Downloads the OSM data of local walking network of defined buffer around defined POI.
}
\examples{
\dontrun{
library(hike4u)

# Buffer of 10000 meters (10 km) around the POI
hiking_routes <- get_hiking_routes(poi, buffer)
}

}
