% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_number_routes.R
\name{get_number_routes}
\alias{get_number_routes}
\title{Get Number of Routes}
\usage{
get_number_routes(final_routes, buffer)
}
\arguments{
\item{final_routes}{the sf dataframe with hiking routes. Created by function "calculate_closeness".}

\item{buffer}{the defined buffer around the POI in meters as a Value. Created by function "create_buffer".}
}
\value{
Returns a text stating the amount of routes within the buffer area in console.
}
\description{
\code{get_number_routes} Prints the number of hiking routes within the buffer area.
}
\examples{
\dontrun{
library(hike4u)

get_number_routes(final_routes, buffer)
}

}
