% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hike4u.R
\name{hike4u}
\alias{hike4u}
\title{hike4u}
\usage{
hike4u(longitude, latitude, buffer, closeness_value)
}
\arguments{
\item{longitude}{the longitude of your location.}

\item{latitude}{the latitude of your location.}

\item{buffer}{the defined buffer around the POI in meters.}

\item{closeness_value}{the defined closeness value. Choose between 1 and the max number of routes in your area. 1 = closest route to your location.}
}
\value{
Returns two maps in you working directory folder and text stating the amount of routes within the buffer area in console.
}
\description{
\code{hike4u} uses your location, given in longitude and latitude, a given buffer size in meters to find hiking routes close to you. It plots two maps, and overview and a satellite map, of the route you have defined by closeness, and saves them to your working directory.
}
\examples{
\dontrun{
library(hike4u)

# Define you location with long=9.93389691622025 and lat=49.79895823510417,
# define the buffer as 10000 meters (10 km) and the closeness value as 1.
hike4u(9.93389691622025, 49.79895823510417, 10000, 1)
}

}
