% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hike4u_ov.R
\name{hike4u_ov}
\alias{hike4u_ov}
\title{Create Overview Map}
\usage{
hike4u_ov(longitude, latitude, buffer, closeness_value)
}
\arguments{
\item{longitude}{the longitude of your location. Will create POI.}

\item{latitude}{the latitude of your location. Will create POI.}

\item{buffer}{a defined buffer around the POI in meters.}

\item{closeness_value}{a defined closeness value. Choose between 1 and the max number of routes in your area. 1 = closest route to your location.}
}
\value{
Returns an overview map in you working directory folder with name: "closest_route_nr_\emph{closeness_value}_overview_map.png.
}
\description{
\code{hike4u_ov} uses your location, given in longitude and latitude, a given buffer size in meters to find hiking routes close to you. It plots an overview map of the route you have defined by closeness, and saves it to your working directory.
}
\examples{
\dontrun{
library(hike4u)

# Define your location with long=9.93389691622025 and lat=49.79895823510417,
# define the buffer as 10000 meters (10 km) and the closeness value as 1
hike4u_ov(9.93389691622025, 49.79895823510417, 10000, 1)
}

}
