% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_routes.R
\name{individual_routes}
\alias{individual_routes}
\title{Extract Individual Routes}
\usage{
individual_routes(hiking_routes)
}
\arguments{
\item{hiking_routes}{the sf dataframe with all downloaded hiking routes. Created by function "get_hiking_routes".}
}
\value{
Returns a filtered sf dataframe "final_routes" with one row/geometry per hiking route.
}
\description{
\code{individual_routes} If one route is split into multiple rows, this function merges the geometries to create a single geometry for the route.
}
\examples{
\dontrun{
library(hike4u)

final_routes <- individual_routes(hiking_routes)
}

}
