% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_mls.R
\name{merge_mls}
\alias{merge_mls}
\title{Merge Multilinestring Geometries}
\usage{
merge_mls(sf_df, column)
}
\arguments{
\item{sf_df}{an sf dataframe with geometries \emph{Multilinestring} and an additional column to merge by.}

\item{column}{the column name by which values to merge the Multilinestring geometries by. Must be passed on as a character string. See example.}
}
\value{
Returns the sf dataframe with one row/geometry per value in defined column.
}
\description{
\code{merge_mls} merges Multilinestring geometries in an sf dataframe according to the value of a given column, to create a single geometry and row.
}
\examples{
\dontrun{
library(hike4u)

# Retrieve the sample dataframe from the package
hiking_routes <- system.file("extdata", "hiking_routes.rds", package = "hike4u")

# Define the sf dataframe "hiking_routes" and the column osm_id within the function
final_routes <- merge_mls(hiking_routes, "osm_id")
}

}
