% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_satellite_map.R
\name{plot_satellite_map}
\alias{plot_satellite_map}
\title{Plot Satellite Map}
\usage{
plot_satellite_map(final_routes, poi, ext, closeness_value)
}
\arguments{
\item{final_routes}{the sf dataframe with hiking routes. Created by function "calculate_closeness".}

\item{poi}{the POI as an sf dataframe with your location as a point. Created by function "create_poi_df".}

\item{ext}{the extent of the map as an sf dataframe. Created by function "calculate_ext".}

\item{closeness_value}{the defined closeness value as a Value. Created by function "def_closeness".}
}
\value{
Returns a PNG file with the satellite map of the chosen route in the working directory. Or "No rows found with the specified closeness value."
}
\description{
\code{plot_satellite_map} plots the satellite map of the chosen, by closeness value, route and exports the map as a PNG file to the working directory.
}
\examples{
\dontrun{
library(hike4u)

plot_satellite_map(final_routes, poi, ext, closeness_value)
}

}
