% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squ_bbox.R
\name{squ_bbox}
\alias{squ_bbox}
\title{Square BBox}
\usage{
squ_bbox(sf_df, column, value, increase)
}
\arguments{
\item{sf_df}{an sf dataframe.}

\item{column}{the column in the sf dataframe with the geometry to create the bbox around. Must be passed on as a character string. See example.}

\item{value}{the value in the column to filter the sf dataframe by. Must be unique.}

\item{increase}{the amount of meters to increase the extent of the bbox by.}
}
\value{
Returns a value of class bbox.
}
\description{
\code{squ_bbox} creates a square bbox around any geometry in a chosen column of an sf dataframe and increases the extent by a chosen amount of meters.
}
\examples{
\dontrun{
library(hike4u)

# Retrieve the sample dataframe from the package
final_routes_cl <- system.file("extdata", "final_routes_cl.rds", package = "hike4u")

# Define the sf dataframe "final_routes_cl" within the function, the column
# name "closeness" as a character string, the closeness value as 1,
# and 1000 meters as the increase value
bbox <- squ_bbox(final_routes_cl, "closeness", 1, 1000)
}

}
