% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_number_routes.R
\name{get_number_routes}
\alias{get_number_routes}
\title{Get Number of Routes}
\usage{
get_number_routes(longitude, latitude, buffer)
}
\arguments{
\item{longitude}{the longitude of your location. Will create POI.}

\item{latitude}{the latitude of your location. Will create POI.}

\item{buffer}{a defined buffer around the POI in meters.}
}
\value{
Returns a message in console stating the amount of routes within the buffer area.
}
\description{
\code{get_number_routes} prints the number of hiking routes within the buffer area.
}
\examples{
\dontrun{
library(hike4u)

# Define your location with long=9.93389691622025 and lat=49.79895823510417,
# and define the buffer as 10000 meters (10 km)
get_number_routes(9.93389691622025, 49.79895823510417, 10000)
}

}
